% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admnrev.R
\docType{data}
\name{admnrev}
\alias{admnrev}
\title{admnrev}
\format{A data.frame with 153 rows and 5 variables:
\describe{
 \item{state}{state postal code}
 \item{year}{85, 90, or 95}
 \item{admnrev}{=1 if admin. revoc. law}
 \item{daysfrst}{days suspended, first offense}
 \item{daysscnd}{days suspended, second offense}
}}
\source{
\url{http://www.cengage.com/c/introductory-econometrics-a-modern-approach-6e-wooldridge}
}
\usage{
data(admnrev)
}
\description{
Data from wooldRidge package loads lazily. Type data(admnrev) into the console.
}
\examples{
 str(admnrev)
}
\keyword{datasets}
