% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/airfare.R
\docType{data}
\name{airfare}
\alias{airfare}
\title{airfare}
\format{A data.frame with 4596 rows and 14 variables:
\describe{
 \item{year}{1997, 1998, 1999, 2000}
 \item{id}{route identifier}
 \item{dist}{distance, in miles}
 \item{passen}{avg. passengers per day}
 \item{fare}{avg. one-way fare, $}
 \item{bmktshr}{fraction market, biggest carrier}
 \item{ldist}{log(distance)}
 \item{y98}{=1 if year == 1998}
 \item{y99}{=1 if year == 1999}
 \item{y00}{=1 if year == 2000}
 \item{lfare}{log(fare)}
 \item{ldistsq}{ldist^2}
 \item{concen}{= bmktshr}
 \item{lpassen}{log(passen)}
}}
\source{
\url{http://www.cengage.com/c/introductory-econometrics-a-modern-approach-6e-wooldridge}
}
\usage{
data(airfare)
}
\description{
Data from wooldRidge package loads lazily. Type data(airfare) into the console.
}
\keyword{datasets}
