% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wooldRidge.R
\docType{data}
\name{apple}
\alias{apple}
\title{apple}
\format{A data.frame with 660 rows and 17 variables:

\describe{
  \item{id}{respondent identifier}
  \item{educ}{years schooling}
  \item{date}{date:  month/day/year}
  \item{state}{home state}
  \item{regprc}{price of regular apples}
  \item{ecoprc}{price of ecolabeled apples}
  \item{inseason}{=1 if interviewed in Nov.}
  \item{hhsize}{household size}
  \item{male}{=1 if male}
  \item{faminc}{family income, thousands}
  \item{age}{in years}
  \item{reglbs}{quantity regular apples, pounds}
  \item{ecolbs}{quantity ecolabeled apples, lbs}
  \item{numlt5}{# in household younger than 5}
  \item{num5_17}{# in household 5 to 17}
  \item{num18_64}{# in household 18 to 64}
  \item{numgt64}{# in household older than 64}
}}
\usage{
apple
}
\description{
apple
}
\keyword{datasets}
