% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/approval.R
\docType{data}
\name{approval}
\alias{approval}
\title{approval}
\format{A tbl_df with 78 rows and 16 variables:}
\source{
\url{http://www.cengage.com/c/introductory-econometrics-a-modern-approach-6e-wooldridge}
}
\usage{
data('approval')
}
\description{
Data from wooldRidge package loads lazily. Type data('approval') into the console.
}
\examples{
 str(approval)
}
\keyword{datasets}
