% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/athlet2.R
\docType{data}
\name{athlet2}
\alias{athlet2}
\title{athlet2}
\format{A data.frame with 30 rows and 10 variables:
\itemize{
 \item dscore. home scr. - vist. scr., 1993
 \item dinstt. diff. in-state tuit., 1994
 \item doutstt. diff. out-state tuit., 1994
 \item htpriv. =1 if home team priv. sch.
 \item vtpriv. =1 if vist. team priv. sch.
 \item dapps. diff. in applications, 1994
 \item htwrd. =1 if home win. record, 1993
 \item vtwrd. =1 if vist. win. record, 1993
 \item dwinrec. htwrd - vtwrd
 \item dpriv. htpriv - vtpriv
}}
\source{
\url{https://www.cengage.com/cgi-wadsworth/course_products_wp.pl?fid=M20b&product_isbn_issn=9781111531041}
}
\usage{
data(athlet2)
}
\description{
Data loads lazily. Type data(athlet2) into the console.
}
\examples{
 str(athlet2)
}
\keyword{datasets}
