% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wooldRidge.R
\docType{data}
\name{attend}
\alias{attend}
\title{attend}
\format{A data.frame with 680 rows and 11 variables:

\describe{
  \item{attend}{classes attended out of 32}
  \item{termGPA}{GPA for term}
  \item{priGPA}{cumulative GPA prior to term}
  \item{ACT}{ACT score}
  \item{final}{final exam score}
  \item{atndrte}{percent classes attended}
  \item{hwrte}{percent homework turned in}
  \item{frosh}{=1 if freshman}
  \item{soph}{=1 if sophomore}
  \item{missed}{number of classes missed}
  \item{stndfnl}{(final - mean)/sd}
}}
\usage{
attend
}
\description{
Data from wooldRidge package loads lazily, simply type attend into the console.
}
\keyword{datasets}
