% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wooldRidge.R
\docType{data}
\name{audit}
\alias{audit}
\title{audit}
\format{A data.frame with 241 rows and 3 variables:

\describe{
  \item{w}{=1 if white app. got job offer}
  \item{b}{=1 if black app. got job offer}
  \item{y}{b - w}
}}
\usage{
audit
}
\description{
audit
}
\keyword{datasets}
