% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wooldRidge.R
\docType{data}
\name{barium}
\alias{barium}
\title{barium}
\format{An object of class \code{data.frame} with 131 rows and 31 columns.}
\usage{
attend
}
\description{
Data from wooldRidge package loads lazily, simply type barium into the console.
}
\keyword{datasets}
