% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bwght.R
\docType{data}
\name{bwght}
\alias{bwght}
\title{bwght}
\format{A data.frame with 1388 rows and 14 variables:
\itemize{
 \item faminc. 1988 family income, $1000s
 \item cigtax. cig. tax in home state, 1988
 \item cigprice. cig. price in home state, 1988
 \item bwght. birth weight, ounces
 \item fatheduc. father's yrs of educ
 \item motheduc. mother's yrs of educ
 \item parity. birth order of child
 \item male. =1 if male child
 \item white. =1 if white
 \item cigs. cigs smked per day while preg
 \item lbwght. log of bwght
 \item bwghtlbs. birth weight, pounds
 \item packs. packs smked per day while preg
 \item lfaminc. log(faminc)
}}
\source{
\url{http://www.cengage.com/cgi-wadsworth/course_products_wp.pl?fid=M20b&product_isbn_issn=9781111531041}
}
\usage{
data(bwght)
}
\description{
Data loads lazily. Type data(bwght) into the console.
}
\examples{
 str(bwght)
}
\keyword{datasets}
