% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bwght2.R
\docType{data}
\name{bwght2}
\alias{bwght2}
\title{bwght2}
\format{A data.frame with 1832 rows and 23 variables:
\describe{
 \item{mage}{mother's age, years}
 \item{meduc}{mother's educ, years}
 \item{monpre}{month prenatal care began}
 \item{npvis}{total number of prenatal visits}
 \item{fage}{father's age, years}
 \item{feduc}{father's educ, years}
 \item{bwght}{birth weight, grams}
 \item{omaps}{one minute apgar score}
 \item{fmaps}{five minute apgar score}
 \item{cigs}{avg cigarettes per day}
 \item{drink}{avg drinks per week}
 \item{lbw}{=1 if bwght <= 2000}
 \item{vlbw}{=1 if bwght <= 1500}
 \item{male}{=1 if baby male}
 \item{mwhte}{=1 if mother white}
 \item{mblck}{=1 if mother black}
 \item{moth}{=1 if mother is other}
 \item{fwhte}{=1 if father white}
 \item{fblck}{=1 if father black}
 \item{foth}{=1 if father is other}
 \item{lbwght}{log(bwght)}
 \item{magesq}{mage^2}
 \item{npvissq}{npvis^2}
}}
\source{
\url{http://www.cengage.com/c/introductory-econometrics-a-modern-approach-6e-wooldridge}
}
\usage{
data(bwght2)
}
\description{
Data from wooldRidge package loads lazily. Type data(bwght2) into the console.
}
\examples{
 str(bwght2)
}
\keyword{datasets}
