% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catholic.R
\docType{data}
\name{catholic}
\alias{catholic}
\title{catholic}
\format{A tbl_df with 7430 rows and 13 variables:
\describe{
 \item{id}{Description not provided}
 \item{read12}{Description not provided}
 \item{math12}{Description not provided}
 \item{female}{Description not provided}
 \item{asian}{Description not provided}
 \item{hispan}{Description not provided}
 \item{black}{Description not provided}
 \item{motheduc}{Description not provided}
 \item{fatheduc}{Description not provided}
 \item{lfaminc}{Description not provided}
 \item{hsgrad}{Description not provided}
 \item{cathhs}{Description not provided}
 \item{parcath}{Description not provided}
}}
\source{
\url{http://www.cengage.com/c/introductory-econometrics-a-modern-approach-6e-wooldridge}
}
\usage{
data('catholic')
}
\description{
Data from wooldRidge package loads lazily. Type data('catholic') into the console.
}
\examples{
 str(catholic)
}
\keyword{datasets}
