% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/census2000.R
\docType{data}
\name{census2000}
\alias{census2000}
\title{census2000}
\format{A data.frame with 29501 observations on 6 variables:
\itemize{
 \item{state}{Description not provided}
 \item{puma}{Description not provided}
 \item{educ}{Description not provided}
 \item{lweekinc}{Description not provided}
 \item{exper}{Description not provided}
 \item{expersq}{Description not provided}
}}
\source{
\url{http://www.cengage.com/c/introductory-econometrics-a-modern-approach-6e-wooldridge}
}
\usage{
data('census2000')
}
\description{
Wooldridge Source: Obtained from the United States Census Bureau by Professor Alberto Abadie of the Harvard Kennedy School of Government. Professor Abadie kindly provided the data. Data loads lazily.
}
\section{Used in Text}{
  pages 452-453
}

\examples{
 str(census2000)
}
\keyword{datasets}
