% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ceosal2.R
\docType{data}
\name{ceosal2}
\alias{ceosal2}
\title{ceosal2}
\format{A data.frame with 177 rows and 15 variables:
\itemize{
 \item salary. 1990 compensation, $1000s
 \item age. in years
 \item college. =1 if attended college
 \item grad. =1 if attended graduate school
 \item comten. years with company
 \item ceoten. years as ceo with company
 \item sales. 1990 firm sales, millions
 \item profits. 1990 profits, millions
 \item mktval. market value, end 1990, mills.
 \item lsalary. log(salary)
 \item lsales. log(sales)
 \item lmktval. log(mktval)
 \item comtensq. comten^2
 \item ceotensq. ceoten^2
 \item profmarg. profits as percent of sales
}}
\source{
\url{http://www.cengage.com/cgi-wadsworth/course_products_wp.pl?fid=M20b&product_isbn_issn=9781111531041}
}
\usage{
data(ceosal2)
}
\description{
Data loads lazily. Type data(ceosal2) into the console.
}
\examples{
 str(ceosal2)
}
\keyword{datasets}
