% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/charity.R
\docType{data}
\name{charity}
\alias{charity}
\title{charity}
\format{A data.frame with 4268 rows and 8 variables:
\describe{
 \item{respond}{=1 if responded with gift}
 \item{gift}{amount of gift, Dutch guilders}
 \item{resplast}{=1 if responded to most recent mailing}
 \item{weekslast}{number of weeks since last response}
 \item{propresp}{response rate to mailings}
 \item{mailsyear}{number of mailings per year}
 \item{giftlast}{amount of most recent gift}
 \item{avggift}{average of past gifts}
}}
\source{
\url{http://www.cengage.com/c/introductory-econometrics-a-modern-approach-6e-wooldridge}
}
\usage{
data(charity)
}
\description{
Data from wooldRidge package loads lazily. Type data(charity) into the console.
}
\examples{
 str(charity)
}
\keyword{datasets}
