% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countymurders.R
\docType{data}
\name{countymurders}
\alias{countymurders}
\title{countymurders}
\format{A tbl_df with 37349 rows and 20 variables:}
\source{
\url{http://www.cengage.com/c/introductory-econometrics-a-modern-approach-6e-wooldridge}
}
\usage{
data('countymurders')
}
\description{
Data from wooldRidge package loads lazily. Type data('countymurders') into the console.
}
\examples{
 str(countymurders)
}
\keyword{datasets}
