% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cps91.R
\docType{data}
\name{cps91}
\alias{cps91}
\title{cps91}
\format{A data.frame with 5634 rows and 24 variables:
\itemize{
 \item husage. husband's age
 \item husunion. =1 if hus. in union
 \item husearns. hus. weekly earns
 \item huseduc. husband's yrs schooling
 \item husblck. =1 if hus. black
 \item hushisp. =1 if hus. hispanic
 \item hushrs. hus. weekly hours
 \item kidge6. =1 if have child >= 6
 \item earns. wife's weekly earnings
 \item age. wife's age
 \item black. =1 if wife black
 \item educ. wife's yrs schooling
 \item hispanic. =1 if wife hispanic
 \item union. =1 if wife in union
 \item faminc. annual family income
 \item husexp. huseduc - husage - 6
 \item exper. age - educ - 6
 \item kidlt6. =1 if have child < 6
 \item hours. wife's weekly hours
 \item expersq. exper^2
 \item nwifeinc. non-wife inc, $1000s
 \item inlf. =1 if wife in labor force
 \item hrwage. earns/hours
 \item lwage. log(hrwage)
}}
\source{
\url{https://www.cengage.com/cgi-wadsworth/course_products_wp.pl?fid=M20b&product_isbn_issn=9781111531041}
}
\usage{
data(cps91)
}
\description{
Data loads lazily. Type data(cps91) into the console.
}
\examples{
 str(cps91)
}
\keyword{datasets}
