% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wooldRidge.R
\docType{data}
\name{discrim}
\alias{discrim}
\title{discrim}
\format{A data.frame with 410 rows and 37 variables:

\describe{
  \item{psoda}{price of medium soda, 1st wave}
  \item{pfries}{price of small fries, 1st wave}
  \item{pentree}{price entree (burger or chicken), 1st wave}
  \item{wagest}{starting wage, 1st wave}
  \item{nmgrs}{number of managers, 1st wave}
  \item{nregs}{number of registers, 1st wave}
  \item{hrsopen}{hours open, 1st wave}
  \item{emp}{number of employees, 1st wave}
  \item{psoda2}{price of medium soday, 2nd wave}
  \item{pfries2}{price of small fries, 2nd wave}
  \item{pentree2}{price entree, 2nd wave}
  \item{wagest2}{starting wage, 2nd wave}
  \item{nmgrs2}{number of managers, 2nd wave}
  \item{nregs2}{number of registers, 2nd wave}
  \item{hrsopen2}{hours open, 2nd wave}
  \item{emp2}{number of employees, 2nd wave}
  \item{compown}{=1 if company owned}
  \item{chain}{BK = 1, KFC = 2, Roy Rogers = 3, Wendy's = 4}
  \item{density}{population density, town}
  \item{crmrte}{crime rate, town}
  \item{state}{NJ = 1, PA = 2}
  \item{prpblck}{proportion black, zipcode}
  \item{prppov}{proportion in poverty, zipcode}
  \item{prpncar}{proportion no car, zipcode}
  \item{hseval}{median housing value, zipcode}
  \item{nstores}{number of stores, zipcode}
  \item{income}{median family income, zipcode}
  \item{county}{county label}
  \item{lpsoda}{log(psoda)}
  \item{lpfries}{log(pfries)}
  \item{lhseval}{log(hseval)}
  \item{lincome}{log(income)}
  \item{ldensity}{log(density)}
  \item{NJ}{=1 for New Jersey}
  \item{BK}{=1 if Burger King}
  \item{KFC}{=1 if Kentucky Fried Chicken}
  \item{RR}{=1 if Roy Rogers}
}}
\usage{
attend
}
\description{
Data from wooldRidge package loads lazily, simply type discrim into the console.
}
\keyword{datasets}
