% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/econmath.R
\docType{data}
\name{econmath}
\alias{econmath}
\title{econmath}
\format{A data.frame with 856 rows and 17 variables:
\describe{
 \item{age}{Description not provided}
 \item{work}{Description not provided}
 \item{study}{Description not provided}
 \item{econhs}{Description not provided}
 \item{colgpa}{Description not provided}
 \item{hsgpa}{Description not provided}
 \item{acteng}{Description not provided}
 \item{actmth}{Description not provided}
 \item{act}{Description not provided}
 \item{mathscr}{Description not provided}
 \item{male}{Description not provided}
 \item{calculus}{Description not provided}
 \item{attexc}{Description not provided}
 \item{attgood}{Description not provided}
 \item{fathcoll}{Description not provided}
 \item{mothcoll}{Description not provided}
 \item{score}{Description not provided}
}}
\source{
\url{http://www.cengage.com/c/introductory-econometrics-a-modern-approach-6e-wooldridge}
}
\usage{
data('econmath')
}
\description{
Data from wooldRidge package loads lazily. Type data('econmath') into the console.
}
\examples{
 str(econmath)
}
\keyword{datasets}
