% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elem94_95.R
\docType{data}
\name{elem94_95}
\alias{elem94_95}
\title{elem94_95}
\format{A data.frame with 1848 rows and 14 variables:
\itemize{
 \item distid. district identifier
 \item schid. school identifier
 \item lunch. percent eligible, free lunch
 \item enrol. enrollment
 \item staff. staff per 1000 students
 \item exppp. expenditures per pupil
 \item avgsal. average teacher salary, $
 \item avgben. average teacher non-salary benefits, $
 \item math4. percent passing 4th grade math test
 \item story4. percent passing 4th grade reading test
 \item bs. avgben/avgsal
 \item lavgsal. log(avgsal)
 \item lenrol. log(enrol)
 \item lstaff. log(staff)
}}
\source{
\url{https://www.cengage.com/cgi-wadsworth/course_products_wp.pl?fid=M20b&product_isbn_issn=9781111531041}
}
\usage{
data(elem94_95)
}
\description{
Data loads lazily. Type data(elem94_95) into the console.
}
\examples{
 str(elem94_95)
}
\keyword{datasets}
