% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wooldRidge.R
\docType{data}
\name{engin}
\alias{engin}
\title{engin}
\format{A data.frame with 403 rows and 17 variables:

\describe{
  \item{male}{=1 if male}
  \item{educ}{highest grade completed}
  \item{wage}{monthly salary, Thai baht}
  \item{swage}{starting wage}
  \item{exper}{years on current job}
  \item{pexper}{previous experience}
  \item{lwage}{log(wage)}
  \item{expersq}{exper^2}
  \item{highgrad}{=1 if high school graduate}
  \item{college}{=1 if college graduate}
  \item{grad}{=1 if some graduate school}
  \item{polytech}{=1 if a polytech}
  \item{highdrop}{=1 if no high school degree}
  \item{lswage}{log(swage)}
  \item{pexpersq}{pexper^2}
  \item{mleeduc}{male*educ}
  \item{mleeduc0}{male*(educ - 14)}
}}
\usage{
engin
}
\description{
engin
}
\keyword{datasets}
