% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ezanders.R
\docType{data}
\name{ezanders}
\alias{ezanders}
\title{ezanders}
\format{A data.frame with 108 rows and 25 variables:
\itemize{
 \item month. name of month
 \item uclms. unemployment claims
 \item ez. =1 if enterprise zone
 \item year. 1980 through 1988
 \item y81. =1 if year == 1981
 \item y82. 
 \item y83. 
 \item y84. 
 \item y85. 
 \item y86. 
 \item y87. 
 \item y88. 
 \item luclms. log(uclms)
 \item jan. =1 if month == JAN
 \item feb. 
 \item mar. 
 \item apr. 
 \item may. 
 \item jun. 
 \item jul. 
 \item aug. 
 \item sep. 
 \item oct. 
 \item nov. 
 \item dec. 
}}
\source{
\url{https://www.cengage.com/cgi-wadsworth/course_products_wp.pl?fid=M20b&product_isbn_issn=9781111531041}
}
\usage{
data(ezanders)
}
\description{
Data loads lazily. Type data(ezanders) into the console.
}
\examples{
 str(ezanders)
}
\keyword{datasets}
