% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wooldRidge.R
\docType{data}
\name{fertil2}
\alias{fertil2}
\title{fertil2}
\format{A data.frame with 4361 rows and 27 variables:

\describe{
  \item{mnthborn}{month woman born}
  \item{yearborn}{year woman born}
  \item{age}{age in years}
  \item{electric}{=1 if has electricity}
  \item{radio}{=1 if has radio}
  \item{tv}{=1 if has tv}
  \item{bicycle}{=1 if has bicycle}
  \item{educ}{years of education}
  \item{ceb}{children ever born}
  \item{agefbrth}{age at first birth}
  \item{children}{number of living children}
  \item{knowmeth}{=1 if know about birth control}
  \item{usemeth}{=1 if ever use birth control}
  \item{monthfm}{month of first marriage}
  \item{yearfm}{year of first marriage}
  \item{agefm}{age at first marriage}
  \item{idlnchld}{'ideal' number of children}
  \item{heduc}{husband's years of education}
  \item{agesq}{age^2}
  \item{urban}{=1 if live in urban area}
  \item{urb_educ}{urban*educ}
  \item{spirit}{=1 if religion == spirit}
  \item{protest}{=1 if religion == protestant}
  \item{catholic}{=1 if religion == catholic}
  \item{frsthalf}{=1 if mnthborn <= 6}
  \item{educ0}{=1 if educ == 0}
  \item{evermarr}{=1 if ever married}
}}
\usage{
attend
}
\description{
Data from wooldRidge package loads lazily, simply type fertil2 into the console.
}
\keyword{datasets}
