% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wooldRidge.R
\docType{data}
\name{gpa2}
\alias{gpa2}
\title{gpa2}
\format{A data.frame with 4137 rows and 12 variables:

\describe{
  \item{sat}{combined SAT score}
  \item{tothrs}{total hours through fall semest}
  \item{colgpa}{GPA after fall semester}
  \item{athlete}{=1 if athlete}
  \item{verbmath}{verbal/math SAT score}
  \item{hsize}{size grad. class, 100s}
  \item{hsrank}{rank in grad. class}
  \item{hsperc}{high school percentile, from top}
  \item{female}{=1 if female}
  \item{white}{=1 if white}
  \item{black}{=1 if black}
  \item{hsizesq}{hsize^2}
}}
\usage{
attend
}
\description{
Data from wooldRidge package loads lazily, simply type gpa2 into the console.
}
\keyword{datasets}
