% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hprice2.R
\docType{data}
\name{hprice2}
\alias{hprice2}
\title{hprice2}
\format{A data.frame with 506 rows and 12 variables:
\describe{
 \item{price}{median housing price, $}
 \item{crime}{crimes committed per capita}
 \item{nox}{nit ox concen; parts per 100m}
 \item{rooms}{avg number of rooms}
 \item{dist}{wght dist to 5 employ centers}
 \item{radial}{access. index to rad. hghwys}
 \item{proptax}{property tax per $1000}
 \item{stratio}{average student-teacher ratio}
 \item{lowstat}{perc of people 'lower status'}
 \item{lprice}{log(price)}
 \item{lnox}{log(nox)}
 \item{lproptax}{log(proptax)}
}}
\source{
\url{http://www.cengage.com/c/introductory-econometrics-a-modern-approach-6e-wooldridge}
}
\usage{
data(hprice2)
}
\description{
Data from wooldRidge package loads lazily. Type data(hprice2) into the console.
}
\keyword{datasets}
