% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wooldRidge.R
\docType{data}
\name{hprice3}
\alias{hprice3}
\title{hprice3}
\format{A data.frame with 321 rows and 19 variables:

\describe{
  \item{year}{1978, 1981}
  \item{age}{age of house}
  \item{agesq}{age^2}
  \item{nbh}{neighborhood, 1-6}
  \item{cbd}{dist. to cent. bus. dstrct, ft.}
  \item{inst}{dist. to interstate, ft.}
  \item{linst}{log(inst)}
  \item{price}{selling price}
  \item{rooms}{# rooms in house}
  \item{area}{square footage of house}
  \item{land}{square footage lot}
  \item{baths}{# bathrooms}
  \item{dist}{dist. from house to incin., ft.}
  \item{ldist}{log(dist)}
  \item{lprice}{log(price)}
  \item{y81}{=1 if year = 1981}
  \item{larea}{log(area)}
  \item{lland}{log(land)}
  \item{linstsq}{linst^2}
}}
\usage{
attend
}
\description{
Data from wooldRidge package loads lazily, simply type hprice3 into the console.
}
\keyword{datasets}
