% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hseinv.R
\docType{data}
\name{hseinv}
\alias{hseinv}
\title{hseinv}
\format{A data.frame with 42 rows and 14 variables:
\describe{
 \item{year}{1947-1988}
 \item{inv}{real housing inv, millions $}
 \item{pop}{population, 1000s}
 \item{price}{housing price index; 1982 = 1}
 \item{linv}{log(inv)}
 \item{lpop}{log(pop)}
 \item{lprice}{log(price)}
 \item{t}{time trend: t=1,...,42}
 \item{invpc}{per capita inv: inv/pop}
 \item{linvpc}{log(invpc)}
 \item{lprice_1}{lprice[_n-1]}
 \item{linvpc_1}{linvpc[_n-1]}
 \item{gprice}{lprice - lprice_1}
 \item{ginvpc}{linvpc - linvpc_1}
}}
\source{
\url{http://www.cengage.com/c/introductory-econometrics-a-modern-approach-6e-wooldridge}
}
\usage{
data(hseinv)
}
\description{
Data from wooldRidge package loads lazily. Type data(hseinv) into the console.
}
\examples{
 str(hseinv)
}
\keyword{datasets}
