% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intdef.R
\docType{data}
\name{intdef}
\alias{intdef}
\title{intdef}
\format{A data.frame with 56 rows and 13 variables:
\describe{
 \item{year}{1948 to 2003}
 \item{i3}{3 month T-bill rate}
 \item{inf}{CPI inflation rate}
 \item{rec}{federal receipts, percent GDP}
 \item{out}{federal outlays, percent GDP}
 \item{def}{out - rec}
 \item{i3_1}{i3[_n-1]}
 \item{inf_1}{inf[_n-1]}
 \item{def_1}{def[_n-1]}
 \item{ci3}{i3 - i3_1}
 \item{cinf}{inf - inf_1}
 \item{cdef}{def - def_1}
 \item{y77}{=1 if year >= 1977; change in FY}
}}
\source{
\url{http://www.cengage.com/c/introductory-econometrics-a-modern-approach-6e-wooldridge}
}
\usage{
data(intdef)
}
\description{
Data from wooldRidge package loads lazily. Type data(intdef) into the console.
}
\examples{
 str(intdef)
}
\keyword{datasets}
