% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wooldRidge.R
\docType{data}
\name{inven}
\alias{inven}
\title{inven}
\format{A data.frame with 37 rows and 13 variables:

\describe{
  \item{year}{1959-1995}
  \item{i3}{3 mo. T-bill rate}
  \item{inf}{CPI inflation rate}
  \item{inven}{inventories, billions '92 $}
  \item{gdp}{GDP, billions '92 $}
  \item{r3}{real interest: i3 - inf}
  \item{cinven}{inven - inven[_n-1]}
  \item{cgdp}{gdp - gdp[_n-1]}
  \item{cr3}{r3 - r3[_n-1]}
  \item{ci3}{i3 - i3[_n-1]}
  \item{cinf}{inf - inf[_n-1]}
  \item{ginven}{log(inven) - log(inven[_n-1])}
  \item{ggdp}{log(gdp) - log(gdp[_n-1])}
}}
\usage{
attend
}
\description{
Data from wooldRidge package loads lazily, simply type inven into the console.
}
\keyword{datasets}
