% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wooldRidge.R
\docType{data}
\name{jtrain2}
\alias{jtrain2}
\title{jtrain2}
\format{A data.frame with 445 rows and 19 variables:

\describe{
  \item{train}{=1 if assigned to job training}
  \item{age}{age in 1977}
  \item{educ}{years of education}
  \item{black}{=1 if black}
  \item{hisp}{=1 if Hispanic}
  \item{married}{=1 if married}
  \item{nodegree}{=1 if no high school degree}
  \item{mosinex}{# mnths prior to 1/78 in expmnt}
  \item{re74}{real earns., 1974, $1000s}
  \item{re75}{real earns., 1975, $1000s}
  \item{re78}{real earns., 1978, $1000s}
  \item{unem74}{=1 if unem. all of 1974}
  \item{unem75}{=1 if unem. all of 1975}
  \item{unem78}{=1 if unem. all of 1978}
  \item{lre74}{log(re74); zero if re74 == 0}
  \item{lre75}{log(re75); zero if re75 == 0}
  \item{lre78}{log(re78); zero if re78 == 0}
  \item{agesq}{age^2}
  \item{mostrn}{months in training}
}}
\usage{
jtrain2
}
\description{
jtrain2
}
\keyword{datasets}
