% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lawsch85.R
\docType{data}
\name{lawsch85}
\alias{lawsch85}
\title{lawsch85}
\format{A data.frame with 156 rows and 21 variables:
\describe{
 \item{rank}{law school ranking}
 \item{salary}{median starting salary}
 \item{cost}{law school cost}
 \item{LSAT}{median LSAT score}
 \item{GPA}{median college GPA}
 \item{libvol}{no. volumes in lib., 1000s}
 \item{faculty}{no. of faculty}
 \item{age}{age of law sch., years}
 \item{clsize}{size of entering class}
 \item{north}{=1 if law sch in north}
 \item{south}{=1 if law sch in south}
 \item{east}{=1 if law sch in east}
 \item{west}{=1 if law sch in west}
 \item{lsalary}{log(salary)}
 \item{studfac}{student-faculty ratio}
 \item{top10}{=1 if ranked in top 10}
 \item{r11_25}{=1 if ranked 11-25}
 \item{r26_40}{=1 if ranked 26-40}
 \item{r41_60}{=1 if ranked 41-60}
 \item{llibvol}{log(libvol)}
 \item{lcost}{log(cost)}
}}
\source{
\url{http://www.cengage.com/c/introductory-econometrics-a-modern-approach-6e-wooldridge}
}
\usage{
data(lawsch85)
}
\description{
Data from wooldRidge package loads lazily. Type data(lawsch85) into the console.
}
\keyword{datasets}
