% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loanapp.R
\docType{data}
\name{loanapp}
\alias{loanapp}
\title{loanapp}
\format{A data.frame with 1989 rows and 59 variables:
\describe{
 \item{occ}{occupancy}
 \item{loanamt}{loan amt in thousands}
 \item{action}{type of action taken}
 \item{msa}{msa number of property}
 \item{suffolk}{=1 if property in suffolk co.}
 \item{appinc}{applicant income, $1000s}
 \item{typur}{type of purchaser of loan}
 \item{unit}{number of units in property}
 \item{married}{=1 if applicant married}
 \item{dep}{number of dependents}
 \item{emp}{years employed in line of work}
 \item{yjob}{years at this job}
 \item{self}{=1 if self employed}
 \item{atotinc}{total monthly income}
 \item{cototinc}{coapp total monthly income}
 \item{hexp}{propose housing expense}
 \item{price}{purchase price}
 \item{other}{other financing, $1000s}
 \item{liq}{liquid assets}
 \item{rep}{no. of credit reports}
 \item{gdlin}{credit history meets guidelines}
 \item{lines}{no. of credit lines on reports}
 \item{mortg}{credit history on mortgage paym}
 \item{cons}{credit history on consumer stuf}
 \item{pubrec}{=1 if filed bankruptcy}
 \item{hrat}{housing exp, percent total inc}
 \item{obrat}{other oblgs,  percent total inc}
 \item{fixadj}{fixed or adjustable rate?}
 \item{term}{term of loan in months}
 \item{apr}{appraised value}
 \item{prop}{type of property}
 \item{inss}{PMI sought}
 \item{inson}{PMI approved}
 \item{gift}{gift as down payment}
 \item{cosign}{is there a cosigner}
 \item{unver}{unverifiable info}
 \item{review}{number of times reviewed}
 \item{netw}{net worth}
 \item{unem}{unemployment rate by industry}
 \item{min30}{=1 if minority pop. > 30percent}
 \item{bd}{=1 if boarded-up val > MSA med}
 \item{mi}{=1 if tract inc > MSA median}
 \item{old}{=1 if applic age > MSA median}
 \item{vr}{=1 if tract vac rte > MSA med}
 \item{sch}{=1 if > 12 years schooling}
 \item{black}{=1 if applicant black}
 \item{hispan}{=1 if applicant Hispanic}
 \item{male}{=1 if applicant male}
 \item{reject}{=1 if action == 3}
 \item{approve}{=1 if action == 1 or 2}
 \item{mortno}{no mortgage history}
 \item{mortperf}{no late mort. payments}
 \item{mortlat1}{one or two late payments}
 \item{mortlat2}{> 2 late payments}
 \item{chist}{=0 if accnts deliq. >= 60 days}
 \item{multi}{=1 if two or more units}
 \item{loanprc}{amt/price}
 \item{thick}{=1 if rep > 2}
 \item{white}{=1 if applicant white}
}}
\source{
\url{http://www.cengage.com/c/introductory-econometrics-a-modern-approach-6e-wooldridge}
}
\usage{
data(loanapp)
}
\description{
Data from wooldRidge package loads lazily. Type data(loanapp) into the console.
}
\keyword{datasets}
