% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wooldRidge.R
\docType{data}
\name{lowbrth}
\alias{lowbrth}
\title{lowbrth}
\format{An object of class \code{data.frame} with 100 rows and 36 columns.}
\usage{
attend
}
\description{
Data from wooldRidge package loads lazily, simply type lowbrth into the console.
}
\keyword{datasets}
