% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wooldRidge.R
\docType{data}
\name{mathpnl}
\alias{mathpnl}
\title{mathpnl}
\format{An object of class \code{data.frame} with 3850 rows and 52 columns.}
\usage{
attend
}
\description{
Data from wooldRidge package loads lazily, simply type mathpnl into the console.
}
\keyword{datasets}
