% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wooldRidge.R
\docType{data}
\name{meap00_01}
\alias{meap00_01}
\title{meap00_01}
\format{An object of class \code{data.frame} with 1692 rows and 9 columns.}
\usage{
attend
}
\description{
Data from wooldRidge package loads lazily, simply type meap00_01 into the console.
}
\keyword{datasets}
