% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meap01.R
\docType{data}
\name{meap01}
\alias{meap01}
\title{meap01}
\format{A data.frame with 1823 rows and 11 variables:
\itemize{
 \item dcode. district code
 \item bcode. building code
 \item math4. percent students satisfactory, 4th grade math
 \item read4. percent students satisfactory, 4th grade reading
 \item lunch. percent students eligible for free or reduced lunch
 \item enroll. school enrollment
 \item expend. total spending, $
 \item exppp. expenditures per pupil: expend/enroll
 \item lenroll. log(enroll)
 \item lexpend. log(expend)
 \item lexppp. log(exppp)
}}
\source{
\url{https://www.cengage.com/cgi-wadsworth/course_products_wp.pl?fid=M20b&product_isbn_issn=9781111531041}
}
\usage{
data(meap01)
}
\description{
Data loads lazily. Type data(meap01) into the console.
}
\examples{
 str(meap01)
}
\keyword{datasets}
