% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wooldRidge.R
\docType{data}
\name{meap93}
\alias{meap93}
\title{meap93}
\format{A data.frame with 408 rows and 17 variables:

\describe{
  \item{lnchprg}{perc of studs in sch lnch prog}
  \item{enroll}{school enrollment}
  \item{staff}{staff per 1000 students}
  \item{expend}{expend. per stud, $}
  \item{salary}{avg. teacher salary, $}
  \item{benefits}{avg. teacher benefits, $}
  \item{droprate}{school dropout rate, perc}
  \item{gradrate}{school graduation rate, perc}
  \item{math10}{perc studs passing MEAP math}
  \item{sci11}{perc studs passing MEAP science}
  \item{totcomp}{salary + benefits}
  \item{ltotcomp}{log(totcomp)}
  \item{lexpend}{log of expend}
  \item{lenroll}{log(enroll)}
  \item{lstaff}{log(staff)}
  \item{bensal}{benefits/salary}
  \item{lsalary}{log(salary)}
}}
\usage{
attend
}
\description{
Data from wooldRidge package loads lazily, simply type meap93 into the console.
}
\keyword{datasets}
