% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meapsingle.R
\docType{data}
\name{meapsingle}
\alias{meapsingle}
\title{meapsingle}
\format{A data.frame with 229 rows and 18 variables:
\describe{
 \item{dcode}{Description not provided}
 \item{bcode}{Description not provided}
 \item{math4}{Description not provided}
 \item{read4}{Description not provided}
 \item{enroll}{Description not provided}
 \item{exppp}{Description not provided}
 \item{free}{Description not provided}
 \item{reduced}{Description not provided}
 \item{lunch}{Description not provided}
 \item{medinc}{Description not provided}
 \item{totchild}{Description not provided}
 \item{married}{Description not provided}
 \item{single}{Description not provided}
 \item{pctsgle}{Description not provided}
 \item{zipcode}{Description not provided}
 \item{lenroll}{Description not provided}
 \item{lexppp}{Description not provided}
 \item{lmedinc}{Description not provided}
}}
\source{
\url{http://www.cengage.com/c/introductory-econometrics-a-modern-approach-6e-wooldridge}
}
\usage{
data('meapsingle')
}
\description{
Data from wooldRidge package loads lazily. Type data('meapsingle') into the console.
}
\examples{
 str(meapsingle)
}
\keyword{datasets}
