% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wooldRidge.R
\docType{data}
\name{minwage}
\alias{minwage}
\title{minwage}
\format{An object of class \code{data.frame} with 612 rows and 58 columns.}
\usage{
attend
}
\description{
Data from wooldRidge package loads lazily, simply type minwage into the console.
}
\keyword{datasets}
