% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wooldRidge.R
\docType{data}
\name{murder}
\alias{murder}
\title{murder}
\format{A data.frame with 153 rows and 13 variables:

\describe{
  \item{id}{state identifier}
  \item{state}{postal code}
  \item{year}{87, 90, or 93}
  \item{mrdrte}{murders per 100,000 people}
  \item{exec}{total executions, past 3 years}
  \item{unem}{annual unem. rate}
  \item{d90}{=1 if year == 90}
  \item{d93}{=1 if year == 93}
  \item{cmrdrte}{mrdrte - mrdrte[_n-1]}
  \item{cexec}{exec - exec[_n-1]}
  \item{cunem}{unem - unem[_n-1]}
  \item{cexec_1}{cexec[_n-1]}
  \item{cunem_1}{cunem[_n-1]}
}}
\usage{
murder
}
\description{
murder
}
\keyword{datasets}
