% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nbasal.R
\docType{data}
\name{nbasal}
\alias{nbasal}
\title{nbasal}
\format{A data.frame with 269 rows and 22 variables:
\describe{
 \item{marr}{=1 if married}
 \item{wage}{annual salary, thousands $}
 \item{exper}{years as professional player}
 \item{age}{age in years}
 \item{coll}{years played in college}
 \item{games}{average games per year}
 \item{minutes}{average minutes per year}
 \item{guard}{=1 if guard}
 \item{forward}{=1 if forward}
 \item{center}{=1 if center}
 \item{points}{points per game}
 \item{rebounds}{rebounds per game}
 \item{assists}{assists per game}
 \item{draft}{draft number}
 \item{allstar}{=1 if ever all star}
 \item{avgmin}{minutes per game}
 \item{lwage}{log(wage)}
 \item{black}{=1 if black}
 \item{children}{=1 if has children}
 \item{expersq}{exper^2}
 \item{agesq}{age^2}
 \item{marrblck}{marr*black}
}}
\source{
\url{http://www.cengage.com/c/introductory-econometrics-a-modern-approach-6e-wooldridge}
}
\usage{
data(nbasal)
}
\description{
Data from wooldRidge package loads lazily. Type data(nbasal) into the console.
}
\examples{
 str(nbasal)
}
\keyword{datasets}
