% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/okun.R
\docType{data}
\name{okun}
\alias{okun}
\title{okun}
\format{A data.frame with 47 rows and 4 variables:
\describe{
 \item{year}{1959 through 2005}
 \item{pcrgdp}{percentage change in real GDP}
 \item{unem}{civilian unemployment rate}
 \item{cunem}{unem - unem[_n-1]}
}}
\source{
\url{http://www.cengage.com/c/introductory-econometrics-a-modern-approach-6e-wooldridge}
}
\usage{
data(okun)
}
\description{
Data from wooldRidge package loads lazily. Type data(okun) into the console.
}
\examples{
 str(okun)
}
\keyword{datasets}
