% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openness.R
\docType{data}
\name{openness}
\alias{openness}
\title{openness}
\format{A data.frame with 114 rows and 12 variables:
\itemize{
 \item open. imports as percent GDP, '73-
 \item inf. avg. annual inflation, '73-
 \item pcinc. 1980 per capita inc., U.S. $
 \item land. land area, square miles
 \item oil. =1 if major oil producer
 \item good. =1 if 'good' data
 \item lpcinc. log(pcinc)
 \item lland. log(land)
 \item lopen. log(open)
 \item linf. log(inf)
 \item opendec. open/100
 \item linfdec. log(inf/100)
}}
\source{
\url{https://www.cengage.com/cgi-wadsworth/course_products_wp.pl?fid=M20b&product_isbn_issn=9781111531041}
}
\usage{
data(openness)
}
\description{
Data loads lazily. Type data(openness) into the console.
}
\examples{
 str(openness)
}
\keyword{datasets}
