% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pension.R
\docType{data}
\name{pension}
\alias{pension}
\title{pension}
\format{A data.frame with 194 rows and 19 variables:
\describe{
 \item{id}{family identifier}
 \item{pyears}{years in pension plan}
 \item{prftshr}{=1 if profit sharing plan}
 \item{choice}{=1 if can choose method invest}
 \item{female}{=1 if female}
 \item{married}{=1 if married}
 \item{age}{age in years}
 \item{educ}{highest grade completed}
 \item{finc25}{$15,000 < faminc92 <= $25,000}
 \item{finc35}{$25,000 < faminc92 <= $35,000}
 \item{finc50}{$35,000 < faminc92 <= $50,000}
 \item{finc75}{$50,000 < faminc92 <= $75,000}
 \item{finc100}{$75,000 < faminc92 <= $100,000}
 \item{finc101}{$100,000 < faminc92}
 \item{wealth89}{net worth, 1989, $1000}
 \item{black}{=1 if black}
 \item{stckin89}{=1 if owned stock in 1989}
 \item{irain89}{=1 if had IRA in 1989}
 \item{pctstck}{0=mstbnds,50=mixed,100=mststcks}
}}
\source{
\url{http://www.cengage.com/c/introductory-econometrics-a-modern-approach-6e-wooldridge}
}
\usage{
data(pension)
}
\description{
Data from wooldRidge package loads lazily. Type data(pension) into the console.
}
\examples{
 str(pension)
}
\keyword{datasets}
