% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wooldRidge.R
\docType{data}
\name{phillips}
\alias{phillips}
\title{phillips}
\format{An object of class \code{data.frame} with 56 rows and 7 columns.}
\usage{
attend
}
\description{
Data from wooldRidge package loads lazily, simply type phillips into the console.
}
\keyword{datasets}
