% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdchem.R
\docType{data}
\name{rdchem}
\alias{rdchem}
\title{rdchem}
\format{A data.frame with 32 rows and 8 variables:
\itemize{
 \item rd. R&D spending, millions
 \item sales. firm sales, millions
 \item profits. profits, millions
 \item rdintens. rd as percent of sales
 \item profmarg. profits as percent of sales
 \item salessq. sales^2
 \item lsales. log(sales)
 \item lrd. log(rd)
}}
\source{
\url{http://www.cengage.com/cgi-wadsworth/course_products_wp.pl?fid=M20b&product_isbn_issn=9781111531041}
}
\usage{
data(rdchem)
}
\description{
Data loads lazily. Type data(rdchem) into the console.
}
\examples{
 str(rdchem)
}
\keyword{datasets}
