% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdtelec.R
\docType{data}
\name{rdtelec}
\alias{rdtelec}
\title{rdtelec}
\format{A data.frame with 29 rows and 6 variables:
\describe{
 \item{rd}{R&D spending, millions $}
 \item{sales}{firm sales, millions $}
 \item{rdintens}{rd as percent of sales}
 \item{lrd}{log(rd)}
 \item{lsales}{log(sales)}
 \item{salessq}{sales^2}
}}
\source{
\url{http://www.cengage.com/c/introductory-econometrics-a-modern-approach-6e-wooldridge}
}
\usage{
data(rdtelec)
}
\description{
Data from wooldRidge package loads lazily. Type data(rdtelec) into the console.
}
\keyword{datasets}
