% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wooldRidge.R
\docType{data}
\name{return}
\alias{return}
\title{return}
\format{An object of class \code{data.frame} with 142 rows and 12 columns.}
\usage{
attend
}
\description{
Data from wooldRidge package loads lazily, simply type return into the console.
}
\keyword{datasets}
