% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wooldRidge.R
\docType{data}
\name{saving}
\alias{saving}
\title{saving}
\format{A data.frame with 100 rows and 7 variables:

\describe{
  \item{sav}{annual savings, $}
  \item{inc}{annual income, $}
  \item{size}{family size}
  \item{educ}{years educ, household head}
  \item{age}{age of household head}
  \item{black}{=1 if household head is black}
  \item{cons}{annual consumption, $}
}}
\usage{
attend
}
\description{
Data from wooldRidge package loads lazily, simply type saving into the console.
}
\keyword{datasets}
