% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slp75_81.R
\docType{data}
\name{slp75_81}
\alias{slp75_81}
\title{slp75_81}
\format{A data.frame with 239 observations on 20 variables:
\itemize{
 \item \strong{age75:} age in 1975
 \item \strong{educ75:} years educ in '75
 \item \strong{educ81:} years educ in '81
 \item \strong{gdhlth75:} = 1 if good hlth in '75
 \item \strong{gdhlth81:} =1 if good hlth in '81
 \item \strong{male:} =1 if male
 \item \strong{marr75:} = 1 if married in '75
 \item \strong{marr81:} =1 if married in '81
 \item \strong{slpnap75:} mins slp wk, inc naps, '75
 \item \strong{slpnap81:} mins slp wk, inc naps, '81
 \item \strong{totwrk75:} minutes worked per week, '75
 \item \strong{totwrk81:} minutes worked per week, '81
 \item \strong{yngkid75:} = 1 if child < 3, '75
 \item \strong{yngkid81:} =1 if child < 3, '81
 \item \strong{ceduc:} change in educ
 \item \strong{cgdhlth:} change in gdhlth
 \item \strong{cmarr:} change in marr
 \item \strong{cslpnap:} change in slpnap
 \item \strong{ctotwrk:} change in totwrk
 \item \strong{cyngkid:} change in yngkid
}}
\source{
\url{https://www.cengage.com/cgi-wadsworth/course_products_wp.pl?fid=M20b&product_isbn_issn=9781111531041}
}
\usage{
data('slp75_81')
}
\description{
Wooldridge Source: See SLEEP75.RAW Data loads lazily.
}
\section{Used in Text}{
 pages 463-464
}

\examples{
 str(slp75_81)
}
\keyword{datasets}
