% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smoke.R
\docType{data}
\name{smoke}
\alias{smoke}
\title{smoke}
\format{A data.frame with 807 rows and 10 variables:
\describe{
 \item{educ}{years of schooling}
 \item{cigpric}{state cig. price, cents/pack}
 \item{white}{=1 if white}
 \item{age}{in years}
 \item{income}{annual income, $}
 \item{cigs}{cigs. smoked per day}
 \item{restaurn}{=1 if rest. smk. restrictions}
 \item{lincome}{log(income)}
 \item{agesq}{age^2}
 \item{lcigpric}{log(cigprice)}
}}
\source{
\url{http://www.cengage.com/c/introductory-econometrics-a-modern-approach-6e-wooldridge}
}
\usage{
data(smoke)
}
\description{
Data from wooldRidge package loads lazily. Type data(smoke) into the console.
}
\examples{
 str(smoke)
}
\keyword{datasets}
