% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wooldRidge.R
\docType{data}
\name{twoyear}
\alias{twoyear}
\title{twoyear}
\format{An object of class \code{data.frame} with 6763 rows and 23 columns.}
\usage{
attend
}
\description{
Data from wooldRidge package loads lazily, simply type twoyear into the console.
}
\keyword{datasets}
