% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wooldRidge.R
\docType{data}
\name{vote1}
\alias{vote1}
\title{vote1}
\format{An object of class \code{data.frame} with 173 rows and 10 columns.}
\usage{
attend
}
\description{
Data from wooldRidge package loads lazily, simply type vote1 into the console.
}
\keyword{datasets}
