% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wooldRidge.R
\docType{data}
\name{vote2}
\alias{vote2}
\title{vote2}
\format{An object of class \code{data.frame} with 186 rows and 26 columns.}
\usage{
attend
}
\description{
Data from wooldRidge package loads lazily, simply type vote2 into the console.
}
\keyword{datasets}
