% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/voucher.R
\docType{data}
\name{voucher}
\alias{voucher}
\title{voucher}
\format{A data.frame with 990 rows and 19 variables:
\describe{
 \item{studyid}{student identifier}
 \item{black}{= 1 if African-American}
 \item{hispanic}{= 1 if Hispanic}
 \item{female}{= 1 if female}
 \item{appyear}{year of first application: 90 to 93}
 \item{mnce}{math NCE test score, 1994}
 \item{select}{= 1 if ever selected to attend choice school}
 \item{choice}{= 1 if attending choice school, 1994}
 \item{selectyrs}{years selected to attend choice school}
 \item{choiceyrs}{years attended choice school}
 \item{mnce90}{mnce in 1990}
 \item{selectyrs1}{= 1 if selectyrs == 1}
 \item{selectyrs2}{= 1 if selectyrs == 2}
 \item{selectyrs3}{= 1 if selectyrs == 3}
 \item{selectyrs4}{= 1 if selectyrs == 4}
 \item{choiceyrs1}{= 1 if choiceyrs == 1}
 \item{choiceyrs2}{= 1 if choiceyrs == 2}
 \item{choiceyrs3}{= 1 if choiceyrs == 3}
 \item{choiceyrs4}{= 1 if choiceyrs == 4}
}}
\source{
\url{http://www.cengage.com/c/introductory-econometrics-a-modern-approach-6e-wooldridge}
}
\usage{
data(voucher)
}
\description{
Data from wooldRidge package loads lazily. Type data(voucher) into the console.
}
\keyword{datasets}
