% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wooldRidge.R
\docType{data}
\name{wagepan}
\alias{wagepan}
\title{wagepan}
\format{A data.frame with 4360 rows and 44 variables:

\describe{
  \item{nr}{person identifier}
  \item{year}{1980 to 1987}
  \item{agric}{=1 if in agriculture}
  \item{black}{=1 if black}
  \item{bus}{}
  \item{construc}{=1 if in construction}
  \item{ent}{}
  \item{exper}{labor mkt experience}
  \item{fin}{}
  \item{hisp}{=1 if Hispanic}
  \item{poorhlth}{=1 if in poor health}
  \item{hours}{annual hours worked}
  \item{manuf}{=1 if in manufacturing}
  \item{married}{=1 if married}
  \item{min}{}
  \item{nrthcen}{=1 if north central}
  \item{nrtheast}{=1 if north east}
  \item{occ1}{}
  \item{occ2}{}
  \item{occ3}{}
  \item{occ4}{}
  \item{occ5}{}
  \item{occ6}{}
  \item{occ7}{}
  \item{occ8}{}
  \item{occ9}{}
  \item{per}{}
  \item{pro}{}
  \item{pub}{}
  \item{rur}{}
  \item{south}{=1 if south}
  \item{educ}{years of schooling}
  \item{tra}{}
  \item{trad}{}
  \item{union}{=1 if in union}
  \item{lwage}{log(wage)}
  \item{d81}{=1 if year == 1981}
  \item{d82}{}
  \item{d83}{}
  \item{d84}{}
  \item{d85}{}
  \item{d86}{}
  \item{d87}{}
  \item{expersq}{exper^2}
}}
\usage{
attend
}
\description{
Data from wooldRidge package loads lazily, simply type wagepan into the console.
}
\keyword{datasets}
