% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/approval.R
\docType{data}
\name{approval}
\alias{approval}
\title{approval}
\format{A data.frame with 78 rows and 16 variables:
\describe{
 \item{id}{Description not provided}
 \item{month}{Description not provided}
 \item{year}{Description not provided}
 \item{sp500}{Description not provided}
 \item{cpi}{Description not provided}
 \item{cpifood}{Description not provided}
 \item{approve}{Description not provided}
 \item{gasprice}{Description not provided}
 \item{unemploy}{Description not provided}
 \item{katrina}{Description not provided}
 \item{rgasprice}{Description not provided}
 \item{lrgasprice}{Description not provided}
 \item{11-Sep}{Description not provided}
 \item{iraqinvade}{Description not provided}
 \item{lsp500}{Description not provided}
 \item{lcpifood}{Description not provided}
}}
\source{
\url{http://www.cengage.com/c/introductory-econometrics-a-modern-approach-6e-wooldridge}
}
\usage{
data('approval')
}
\description{
Data from wooldRidge package loads lazily. Type data('approval') into the console.
}
\examples{
 str(approval)
}
\keyword{datasets}
