% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/audit.R
\docType{data}
\name{audit}
\alias{audit}
\title{audit}
\format{A data.frame with 241 rows and 3 variables:
\itemize{
 \item w. =1 if white app. got job offer
 \item b. =1 if black app. got job offer
 \item y. b - w
}}
\source{
\url{http://www.cengage.com/cgi-wadsworth/course_products_wp.pl?fid=M20b&product_isbn_issn=9781111531041}
}
\usage{
data(audit)
}
\description{
Data loads lazily. Type data(audit) into the console.
}
\examples{
 str(audit)
}
\keyword{datasets}
