% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/audit.R
\docType{data}
\name{audit}
\alias{audit}
\title{audit}
\format{
A data.frame with 241 observations on 3 variables:
\itemize{
 \item \strong{w:} =1 if white app. got job offer
 \item \strong{b:} =1 if black app. got job offer
 \item \strong{y:} b - w
}
}
\source{
\url{https://www.cengage.com/cgi-wadsworth/course_products_wp.pl?fid=M20b&product_isbn_issn=9781111531041}
}
\usage{
data('audit')
}
\description{
Wooldridge Source: These data come from a 1988 Urban Institute audit study in the Washington, D.C. area. I obtained them from the article “The Urban Institute Audit Studies: Their Methods and Findings,” by James J. Heckman and Peter Siegelman. In Fix, M. and Struyk, R., eds., Clear and Convincing Evidence: Measurement of Discrimination in America. Washington, D.C.: Urban Institute Press, 1993, 187-258. Data loads lazily.
}
\section{Used in Text}{
 pages 776-777, 784, 787
}

\examples{
 str(audit)
}
\keyword{datasets}
