% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beauty.R
\docType{data}
\name{beauty}
\alias{beauty}
\title{beauty}
\format{A data.frame with 1260 rows and 17 variables:
\itemize{
 \item wage. hourly wage
 \item lwage. log(wage)
 \item belavg. =1 if looks <= 2
 \item abvavg. =1 if looks >=4
 \item exper. years of workforce experience
 \item looks. from 1 to 5
 \item union. =1 if union member
 \item goodhlth. =1 if good health
 \item black. =1 if black
 \item female. =1 if female
 \item married. =1 if married
 \item south. =1 if live in south
 \item bigcity. =1 if live in big city
 \item smllcity. =1 if live in small city
 \item service. =1 if service industry
 \item expersq. exper^2
 \item educ. years of schooling
}}
\source{
\url{http://www.cengage.com/cgi-wadsworth/course_products_wp.pl?fid=M20b&product_isbn_issn=9781111531041}
}
\usage{
data(beauty)
}
\description{
Data loads lazily. Type data(beauty) into the console.
}
\examples{
 str(beauty)
}
\keyword{datasets}
