% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beveridge.R
\docType{data}
\name{beveridge}
\alias{beveridge}
\title{beveridge}
\format{A data.frame with 135 rows and 8 variables:
\describe{
 \item{month}{dec 200 through feb 2012}
 \item{urate}{unemployment rate, percent}
 \item{vrate}{vacancy rate, percent}
 \item{t}{linear time trend}
 \item{urate_1}{L.urate}
 \item{vrate_1}{L.vrate}
 \item{curate}{D.urate}
 \item{cvrate}{D.vrate}
}}
\source{
\url{http://www.cengage.com/c/introductory-econometrics-a-modern-approach-6e-wooldridge}
}
\usage{
data(beveridge)
}
\description{
Data from wooldRidge package loads lazily. Type data(beveridge) into the console.
}
\keyword{datasets}
