% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bwght.R
\docType{data}
\name{bwght}
\alias{bwght}
\title{bwght}
\format{
A data.frame with 1388 observations on 14 variables:
\itemize{
 \item \strong{faminc:} 1988 family income, $1000s
 \item \strong{cigtax:} cig. tax in home state, 1988
 \item \strong{cigprice:} cig. price in home state, 1988
 \item \strong{bwght:} birth weight, ounces
 \item \strong{fatheduc:} father's yrs of educ
 \item \strong{motheduc:} mother's yrs of educ
 \item \strong{parity:} birth order of child
 \item \strong{male:} =1 if male child
 \item \strong{white:} =1 if white
 \item \strong{cigs:} cigs smked per day while preg
 \item \strong{lbwght:} log of bwght
 \item \strong{bwghtlbs:} birth weight, pounds
 \item \strong{packs:} packs smked per day while preg
 \item \strong{lfaminc:} log(faminc)
}
}
\source{
\url{https://www.cengage.com/cgi-wadsworth/course_products_wp.pl?fid=M20b&product_isbn_issn=9781111531041}
}
\usage{
data('bwght')
}
\description{
Wooldridge Source: J. Mullahy (1997), “Instrumental-Variable Estimation of Count Data Models: Applications to Models of Cigarette Smoking Behavior,” Review of Economics and Statistics 79, 596-593. Professor Mullahy kindly provided the data. He obtained them from the 1988 National Health Interview Survey. Data loads lazily.
}
\section{Used in Text}{
 pages 18, 61, 110, 151, 165, 178, 184, 187-188, 258-259, 522-523
}

\examples{
 str(bwght)
}
\keyword{datasets}
