% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/campus.R
\docType{data}
\name{campus}
\alias{campus}
\title{campus}
\format{A data.frame with 97 rows and 7 variables:
\describe{
 \item{enroll}{total enrollment}
 \item{priv}{=1 if private college}
 \item{police}{employed officers}
 \item{crime}{total campus crimes}
 \item{lcrime}{log(crime)}
 \item{lenroll}{log(enroll)}
 \item{lpolice}{log(police)}
}}
\source{
\url{http://www.cengage.com/c/introductory-econometrics-a-modern-approach-6e-wooldridge}
}
\usage{
data(campus)
}
\description{
Data from wooldRidge package loads lazily. Type data(campus) into the console.
}
\keyword{datasets}
