% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wooldRidge.R
\docType{data}
\name{campus}
\alias{campus}
\title{campus}
\format{A data.frame with 97 rows and 7 variables:

\describe{
  \item{enroll}{total enrollment}
  \item{priv}{=1 if private college}
  \item{police}{employed officers}
  \item{crime}{total campus crimes}
  \item{lcrime}{log(crime)}
  \item{lenroll}{log(enroll)}
  \item{lpolice}{log(police)}
}}
\usage{
attend
}
\description{
Data from wooldRidge package loads lazily, simply type campus into the console.
}
\keyword{datasets}
