% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catholic.R
\docType{data}
\name{catholic}
\alias{catholic}
\title{catholic}
\format{A tbl_df with 7430 rows and 13 variables:}
\source{
\url{http://www.cengage.com/c/introductory-econometrics-a-modern-approach-6e-wooldridge}
}
\usage{
data('catholic')
}
\description{
Data from wooldRidge package loads lazily. Type data('catholic') into the console.
}
\examples{
 str(catholic)
}
\keyword{datasets}
