% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catholic.R
\docType{data}
\name{catholic}
\alias{catholic}
\title{catholic}
\format{A data.frame with 7430 observations on 13 variables:
\itemize{
 \item{id}{Description not provided}
 \item{read12}{Description not provided}
 \item{math12}{Description not provided}
 \item{female}{Description not provided}
 \item{asian}{Description not provided}
 \item{hispan}{Description not provided}
 \item{black}{Description not provided}
 \item{motheduc}{Description not provided}
 \item{fatheduc}{Description not provided}
 \item{lfaminc}{Description not provided}
 \item{hsgrad}{Description not provided}
 \item{cathhs}{Description not provided}
 \item{parcath}{Description not provided}
}}
\source{
\url{http://www.cengage.com/c/introductory-econometrics-a-modern-approach-6e-wooldridge}
}
\usage{
data('catholic')
}
\description{
Wooldridge Source: Altonji, J.G., T.E. Elder, and C.R. Taber (2005), “An Evaluation of Instrumental Variable Strategies for Estimating the Effects of Catholic Schooling,” Journal of Human Resources 40, 791-821. Professor Elder kindly provided a subset of the data, with some variables stripped away for confidentiality reasons. Data loads lazily.
}
\section{Used in Text}{
 pages 267, 551
}

\examples{
 str(catholic)
}
\keyword{datasets}
