% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/census2000.R
\docType{data}
\name{census2000}
\alias{census2000}
\title{census2000}
\format{A tbl_df with 29501 rows and 6 variables:}
\source{
\url{http://www.cengage.com/c/introductory-econometrics-a-modern-approach-6e-wooldridge}
}
\usage{
data('census2000')
}
\description{
Data from wooldRidge package loads lazily. Type data('census2000') into the console.
}
\examples{
 str(census2000)
}
\keyword{datasets}
