% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wooldRidge.R
\docType{data}
\name{ceosal1}
\alias{ceosal1}
\title{ceosal1}
\format{An object of class \code{data.frame} with 209 rows and 12 columns.}
\usage{
attend
}
\description{
Data from wooldRidge package loads lazily, simply type ceosal1 into the console.
}
\keyword{datasets}
