% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ceosal1.R
\docType{data}
\name{ceosal1}
\alias{ceosal1}
\title{ceosal1}
\format{A data.frame with 209 rows and 12 variables:
\itemize{
 \item salary. 1990 salary, thousands $
 \item pcsalary. percent change salary, 89-90
 \item sales. 1990 firm sales, millions $
 \item roe. return on equity, 88-90 avg
 \item pcroe. percent change roe, 88-90
 \item ros. return on firm's stock, 88-90
 \item indus. =1 if industrial firm
 \item finance. =1 if financial firm
 \item consprod. =1 if consumer product firm
 \item utility. =1 if transport. or utilties
 \item lsalary. natural log of salary
 \item lsales. natural log of sales
}}
\source{
\url{https://www.cengage.com/cgi-wadsworth/course_products_wp.pl?fid=M20b&product_isbn_issn=9781111531041}
}
\usage{
data(ceosal1)
}
\description{
Data loads lazily. Type data(ceosal1) into the console.
}
\examples{
 str(ceosal1)
}
\keyword{datasets}
