% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/charity.R
\docType{data}
\name{charity}
\alias{charity}
\title{charity}
\format{
A data.frame with 4268 observations on 8 variables:
\itemize{
 \item \strong{respond:} =1 if responded with gift
 \item \strong{gift:} amount of gift, Dutch guilders
 \item \strong{resplast:} =1 if responded to most recent mailing
 \item \strong{weekslast:} number of weeks since last response
 \item \strong{propresp:} response rate to mailings
 \item \strong{mailsyear:} number of mailings per year
 \item \strong{giftlast:} amount of most recent gift
 \item \strong{avggift:} average of past gifts
}
}
\source{
\url{https://www.cengage.com/cgi-wadsworth/course_products_wp.pl?fid=M20b&product_isbn_issn=9781111531041}
}
\usage{
data('charity')
}
\description{
Wooldridge Source: P.H. Franses and R. Paap (2001), Quantitative Models in Marketing Research. Cambridge: Cambridge University Press. Professor Franses kindly provided the data. Data loads lazily.
}
\section{Notes}{
 This data set can be used to illustrate probit and Tobit models, and to study the linear approximations to them.

Used in Text: pages 65, 112-113, 266-267, 628
}

\examples{
 str(charity)
}
\keyword{datasets}
