% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corn.R
\docType{data}
\name{corn}
\alias{corn}
\title{corn}
\format{A data.frame with 37 rows and 5 variables:
\itemize{
 \item county. county number
 \item cornhec. corn per hectare
 \item soyhec. soybeans per hectare
 \item cornpix. corn pixels per hectare
 \item soypix. soy pixels per hectare
}}
\source{
\url{https://www.cengage.com/cgi-wadsworth/course_products_wp.pl?fid=M20b&product_isbn_issn=9781111531041}
}
\usage{
data(corn)
}
\description{
Data loads lazily. Type data(corn) into the console.
}
\examples{
 str(corn)
}
\keyword{datasets}
