% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wooldRidge.R
\docType{data}
\name{corn}
\alias{corn}
\title{corn}
\format{A data.frame with 37 rows and 5 variables:

\describe{
  \item{county}{county number}
  \item{cornhec}{corn per hectare}
  \item{soyhec}{soybeans per hectare}
  \item{cornpix}{corn pixels per hectare}
  \item{soypix}{soy pixels per hectare}
}}
\usage{
attend
}
\description{
Data from wooldRidge package loads lazily, simply type corn into the console.
}
\keyword{datasets}
