% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countymurders.R
\docType{data}
\name{countymurders}
\alias{countymurders}
\title{countymurders}
\format{A data.frame with 37349 observations on 20 variables:
\itemize{
 \item{arrests}{Description not provided}
 \item{countyid}{Description not provided}
 \item{density}{Description not provided}
 \item{popul}{Description not provided}
 \item{perc1019}{Description not provided}
 \item{perc2029}{Description not provided}
 \item{percblack}{Description not provided}
 \item{percmale}{Description not provided}
 \item{rpcincmaint}{Description not provided}
 \item{rpcpersinc}{Description not provided}
 \item{rpcunemins}{Description not provided}
 \item{year}{Description not provided}
 \item{murders}{Description not provided}
 \item{murdrate}{Description not provided}
 \item{arrestrate}{Description not provided}
 \item{statefips}{Description not provided}
 \item{countyfips}{Description not provided}
 \item{execs}{Description not provided}
 \item{lpopul}{Description not provided}
 \item{execrate}{Description not provided}
}}
\source{
\url{http://www.cengage.com/c/introductory-econometrics-a-modern-approach-6e-wooldridge}
}
\usage{
data('countymurders')
}
\description{
Wooldridge Source: Compiled by J. Monroe Gamble for a Summer Research Opportunities Program (SROP) at Michigan State University, Summer 2014. Monroe obtained data from the U.S. Census Bureau, the FBI Uniform Crime Reports, and the Death Penalty Information Center. Data loads lazily.
}
\section{Used in Text}{
 pages 16, 58, 431, 457
}

\examples{
 str(countymurders)
}
\keyword{datasets}
