% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/econmath.R
\docType{data}
\name{econmath}
\alias{econmath}
\title{econmath}
\format{A tbl_df with 856 rows and 17 variables:}
\source{
\url{http://www.cengage.com/c/introductory-econometrics-a-modern-approach-6e-wooldridge}
}
\usage{
data('econmath')
}
\description{
Data from wooldRidge package loads lazily. Type data('econmath') into the console.
}
\examples{
 str(econmath)
}
\keyword{datasets}
