% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/econmath.R
\docType{data}
\name{econmath}
\alias{econmath}
\title{econmath}
\format{A data.frame with 856 observations on 17 variables:
\itemize{
 \item \strong{age: }{age in years}
 \item \strong{work: }{hours worked per week}
 \item \strong{study: }{hours studying per week}
 \item \strong{econhs: }{=1 if economics in high school}
 \item \strong{colgpa: }{college GPA, beginning semester}
 \item \strong{hsgpa: }{high school GPA}
 \item \strong{acteng: }{ACT English score}
 \item \strong{actmth: }{ACT math score}
 \item \strong{act: }{ACT composite}
 \item \strong{mathscr: }{math quiz score, 0-10}
 \item \strong{male: }{=1 if male}
 \item \strong{calculus: }{=1 if taken calculus course}
 \item \strong{attexc: }{=1 if past attndce 'excellent'}
 \item \strong{attgood: }{=1 if past attndce 'good'}
 \item \strong{fathcoll: }{=1 if father has BA}
 \item \strong{mothcoll: }{=1 if mother has BA}
 \item \strong{score: }{course score, in percent}
}}
\source{
\url{http://www.cengage.com/c/introductory-econometrics-a-modern-approach-6e-wooldridge}
}
\usage{
data('econmath')
}
\description{
Wooldridge Source: Compiled by Professor Charles Ballard, Michigan State University Department of Economics. Professor Ballard kindly provided the data. Data loads lazily.
}
\section{Used in Text}{
 167, 185
}

\examples{
 str(econmath)
}
\keyword{datasets}
