% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expendshares.R
\docType{data}
\name{expendshares}
\alias{expendshares}
\title{expendshares}
\format{A data.frame with 1519 rows and 13 variables:
\describe{
 \item{sfood}{share of food expenditures (out of total)}
 \item{sfuel}{share of fuel expenditures}
 \item{sclothes}{share of clothing expenditures}
 \item{salcohol}{share of alcohol expenditures}
 \item{stransport}{share of transportation expenditures}
 \item{sother}{share of other expenditures}
 \item{totexpend}{total expenditure, British pounds per week}
 \item{income}{family income, British pounds per week}
 \item{age}{age of household head}
 \item{kids}{number of children: 1 or 2}
 \item{ltotexpend}{log(totexpend)}
 \item{lincome}{log(income)}
 \item{agesq}{age^2}
}}
\source{
\url{http://www.cengage.com/c/introductory-econometrics-a-modern-approach-6e-wooldridge}
}
\usage{
data(expendshares)
}
\description{
Data from wooldRidge package loads lazily. Type data(expendshares) into the console.
}
\examples{
 str(expendshares)
}
\keyword{datasets}
