% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wooldRidge.R
\docType{data}
\name{ezanders}
\alias{ezanders}
\title{ezanders}
\format{A data.frame with 108 rows and 25 variables:

\describe{
  \item{month}{name of month}
  \item{uclms}{unemployment claims}
  \item{ez}{=1 if enterprise zone}
  \item{year}{1980 through 1988}
  \item{y81}{=1 if year == 1981}
  \item{y82}{}
  \item{y83}{}
  \item{y84}{}
  \item{y85}{}
  \item{y86}{}
  \item{y87}{}
  \item{y88}{}
  \item{luclms}{log(uclms)}
  \item{jan}{=1 if month == JAN}
  \item{feb}{}
  \item{mar}{}
  \item{apr}{}
  \item{may}{}
  \item{jun}{}
  \item{jul}{}
  \item{aug}{}
  \item{sep}{}
  \item{oct}{}
  \item{nov}{}
  \item{dec}{}
}}
\usage{
ezanders
}
\description{
ezanders
}
\keyword{datasets}
