% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/happiness.R
\docType{data}
\name{happiness}
\alias{happiness}
\title{happiness}
\format{
A data.frame with 17137 observations on 33 variables:
\itemize{
 \item \strong{year:} gss year for this respondent
 \item \strong{workstat:} work force status
 \item \strong{prestige:} occupational prestige score
 \item \strong{divorce:} ever been divorced or separated
 \item \strong{widowed:} ever been widowed
 \item \strong{educ:} highest year of school completed
 \item \strong{reg16:} region of residence, age 16
 \item \strong{babies:} household members less than 6 yrs old
 \item \strong{preteen:} household members 6 thru 12 yrs old
 \item \strong{teens:} household members 13 thru 17 yrs old
 \item \strong{income:} total family income
 \item \strong{region:} region of interview
 \item \strong{attend:} how often r attends religious services
 \item \strong{happy:} general happiness
 \item \strong{owngun:} =1 if own gun
 \item \strong{tvhours:} hours per day watching tv
 \item \strong{vhappy:} =1 if 'very happy'
 \item \strong{mothfath16:} =1 if live with mother and father at 16
 \item \strong{black:} =1 if black
 \item \strong{gwbush04:} =1 if voted for G.W. Bush in 2004
 \item \strong{female:} =1 if female
 \item \strong{blackfemale:} black*female
 \item \strong{gwbush00:} =1 if voted for G.W. Bush in 2000
 \item \strong{occattend:} =1 if attend is 3, 4, or 5
 \item \strong{regattend:} =1 if attend is 6, 7, or 8
 \item \strong{y94:} =1 if year == 1994
 \item \strong{y96:} 
 \item \strong{y98:} 
 \item \strong{y00:} 
 \item \strong{y02:} 
 \item \strong{y04:} 
 \item \strong{y06:} =1 if year == 2006
 \item \strong{unem10:} =1 if unemployed in last 10 years
}
}
\source{
\url{https://www.cengage.com/cgi-wadsworth/course_products_wp.pl?fid=M20b&product_isbn_issn=9781111531041}
}
\usage{
data('happiness')
}
\description{
Wooldridge  Data loads lazily.
}
\examples{
 str(happiness)
}
\keyword{datasets}
