% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wooldRidge.R
\docType{data}
\name{hprice1}
\alias{hprice1}
\title{hprice1}
\format{A data.frame with 88 rows and 10 variables:

\describe{
  \item{price}{house price, $1000s}
  \item{assess}{assessed value, $1000s}
  \item{bdrms}{number of bdrms}
  \item{lotsize}{size of lot in square feet}
  \item{sqrft}{size of house in square feet}
  \item{colonial}{=1 if home is colonial style}
  \item{lprice}{log(price)}
  \item{lassess}{log(assess}
  \item{llotsize}{log(lotsize)}
  \item{lsqrft}{log(sqrft)}
}}
\usage{
attend
}
\description{
Data from wooldRidge package loads lazily, simply type hprice1 into the console.
}
\keyword{datasets}
