% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meap00_01.R
\docType{data}
\name{meap00_01}
\alias{meap00_01}
\title{meap00_01}
\format{A data.frame with 1692 rows and 9 variables:
\itemize{
 \item dcode. district code
 \item bcode. building code
 \item math4. percent students satisfactory, 4th grade math
 \item read4. percent students satisfactory, 4th grade reading
 \item lunch. percent students eligible for free or reduced lunch
 \item enroll. school enrollment
 \item exppp. expenditures per pupil: expend/enroll
 \item lenroll. log(enroll)
 \item lexppp. log(exppp)
}}
\source{
\url{https://www.cengage.com/cgi-wadsworth/course_products_wp.pl?fid=M20b&product_isbn_issn=9781111531041}
}
\usage{
data(meap00_01)
}
\description{
Data loads lazily. Type data(meap00_01) into the console.
}
\examples{
 str(meap00_01)
}
\keyword{datasets}
