% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wooldRidge.R
\docType{data}
\name{meap01}
\alias{meap01}
\title{meap01}
\format{An object of class \code{data.frame} with 1823 rows and 11 columns.}
\usage{
attend
}
\description{
Data from wooldRidge package loads lazily, simply type meap01 into the console.
}
\keyword{datasets}
