% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meap01.R
\docType{data}
\name{meap01}
\alias{meap01}
\title{meap01}
\format{A data.frame with 1823 rows and 11 variables:
\describe{
 \item{dcode}{district code}
 \item{bcode}{building code}
 \item{math4}{percent students satisfactory, 4th grade math}
 \item{read4}{percent students satisfactory, 4th grade reading}
 \item{lunch}{percent students eligible for free or reduced lunch}
 \item{enroll}{school enrollment}
 \item{expend}{total spending, $}
 \item{exppp}{expenditures per pupil: expend/enroll}
 \item{lenroll}{log(enroll)}
 \item{lexpend}{log(expend)}
 \item{lexppp}{log(exppp)}
}}
\source{
\url{http://www.cengage.com/c/introductory-econometrics-a-modern-approach-6e-wooldridge}
}
\usage{
data(meap01)
}
\description{
Data from wooldRidge package loads lazily. Type data(meap01) into the console.
}
\examples{
 str(meap01)
}
\keyword{datasets}
