% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meapsingle.R
\docType{data}
\name{meapsingle}
\alias{meapsingle}
\title{meapsingle}
\format{A tbl_df with 229 rows and 18 variables:}
\source{
\url{http://www.cengage.com/c/introductory-econometrics-a-modern-approach-6e-wooldridge}
}
\usage{
data('meapsingle')
}
\description{
Data from wooldRidge package loads lazily. Type data('meapsingle') into the console.
}
\examples{
 str(meapsingle)
}
\keyword{datasets}
