% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wooldRidge.R
\docType{data}
\name{okun}
\alias{okun}
\title{okun}
\format{A data.frame with 47 rows and 4 variables:

\describe{
  \item{year}{1959 through 2005}
  \item{pcrgdp}{percentage change in real GDP}
  \item{unem}{civilian unemployment rate}
  \item{cunem}{unem - unem[_n-1]}
}}
\usage{
okun
}
\description{
okun
}
\keyword{datasets}
