% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wooldRidge.R
\docType{data}
\name{openness}
\alias{openness}
\title{openness}
\format{An object of class \code{data.frame} with 114 rows and 12 columns.}
\usage{
attend
}
\description{
Data from wooldRidge package loads lazily, simply type openness into the console.
}
\keyword{datasets}
