% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phillips.R
\docType{data}
\name{phillips}
\alias{phillips}
\title{phillips}
\format{A data.frame with 56 rows and 7 variables:
\describe{
 \item{year}{1948 through 2003}
 \item{unem}{civilian unemployment rate, percent}
 \item{inf}{percentage change in CPI}
 \item{inf_1}{inf[_n-1]}
 \item{unem_1}{unem[_n-1]}
 \item{cinf}{inf - inf_1}
 \item{cunem}{unem - unem_1}
}}
\source{
\url{http://www.cengage.com/c/introductory-econometrics-a-modern-approach-6e-wooldridge}
}
\usage{
data(phillips)
}
\description{
Data from wooldRidge package loads lazily. Type data(phillips) into the console.
}
\examples{
 str(phillips)
}
\keyword{datasets}
