% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdchem.R
\docType{data}
\name{rdchem}
\alias{rdchem}
\title{rdchem}
\format{A data.frame with 32 rows and 8 variables:
\describe{
 \item{rd}{R&D spending, millions}
 \item{sales}{firm sales, millions}
 \item{profits}{profits, millions}
 \item{rdintens}{rd as percent of sales}
 \item{profmarg}{profits as percent of sales}
 \item{salessq}{sales^2}
 \item{lsales}{log(sales)}
 \item{lrd}{log(rd)}
}}
\source{
\url{http://www.cengage.com/c/introductory-econometrics-a-modern-approach-6e-wooldridge}
}
\usage{
data(rdchem)
}
\description{
Data from wooldRidge package loads lazily. Type data(rdchem) into the console.
}
\keyword{datasets}
