% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wooldRidge.R
\docType{data}
\name{rdchem}
\alias{rdchem}
\title{rdchem}
\format{A data.frame with 32 rows and 8 variables:

\describe{
  \item{rd}{R&D spending, millions}
  \item{sales}{firm sales, millions}
  \item{profits}{profits, millions}
  \item{rdintens}{rd as percent of sales}
  \item{profmarg}{profits as percent of sales}
  \item{salessq}{sales^2}
  \item{lsales}{log(sales)}
  \item{lrd}{log(rd)}
}}
\usage{
attend
}
\description{
Data from wooldRidge package loads lazily, simply type rdchem into the console.
}
\keyword{datasets}
