% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wooldRidge.R
\docType{data}
\name{rdtelec}
\alias{rdtelec}
\title{rdtelec}
\format{A data.frame with 29 rows and 6 variables:

\describe{
  \item{rd}{R&D spending, millions $}
  \item{sales}{firm sales, millions $}
  \item{rdintens}{rd as percent of sales}
  \item{lrd}{log(rd)}
  \item{lsales}{log(sales)}
  \item{salessq}{sales^2}
}}
\usage{
rdtelec
}
\description{
rdtelec
}
\keyword{datasets}
