% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/return.R
\docType{data}
\name{return}
\alias{return}
\title{return}
\format{A data.frame with 142 rows and 12 variables:
\itemize{
 \item roe. return on equity, 1990
 \item rok. return on capital, 1990
 \item dkr. debt/capital, 1990
 \item eps. earnings per share, 1990
 \item netinc. net income, 1990 (mills.)
 \item sp90. stock price, end 1990
 \item sp94. stock price, end 1994
 \item salary. CEO salary, 1990 (thous.)
 \item return. percent change s.p., 90-94
 \item lsalary. log(salary)
 \item lsp90. log(sp90)
 \item lnetinc. log(netinc)
}}
\source{
\url{http://www.cengage.com/cgi-wadsworth/course_products_wp.pl?fid=M20b&product_isbn_issn=9781111531041}
}
\usage{
data(return)
}
\description{
Data loads lazily. Type data(return) into the console.
}
\examples{
 str(return)
}
\keyword{datasets}
