% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saving.R
\docType{data}
\name{saving}
\alias{saving}
\title{saving}
\format{A data.frame with 100 rows and 7 variables:
\itemize{
 \item sav. annual savings, $
 \item inc. annual income, $
 \item size. family size
 \item educ. years educ, household head
 \item age. age of household head
 \item black. =1 if household head is black
 \item cons. annual consumption, $
}}
\source{
\url{http://www.cengage.com/cgi-wadsworth/course_products_wp.pl?fid=M20b&product_isbn_issn=9781111531041}
}
\usage{
data(saving)
}
\description{
Data loads lazily. Type data(saving) into the console.
}
\examples{
 str(saving)
}
\keyword{datasets}
