% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saving.R
\docType{data}
\name{saving}
\alias{saving}
\title{saving}
\format{
A data.frame with 100 observations on 7 variables:
\itemize{
 \item \strong{sav:} annual savings, $
 \item \strong{inc:} annual income, $
 \item \strong{size:} family size
 \item \strong{educ:} years educ, household head
 \item \strong{age:} age of household head
 \item \strong{black:} =1 if household head is black
 \item \strong{cons:} annual consumption, $
}
}
\source{
\url{https://www.cengage.com/cgi-wadsworth/course_products_wp.pl?fid=M20b&product_isbn_issn=9781111531041}
}
\usage{
data('saving')
}
\description{
Wooldridge Source: Unknown Data loads lazily.
}
\section{Notes}{
 I remember entering this data set in the late 1980s, and I am pretty sure it came directly from an introductory econometrics text. But so far my search has been fruitless. If anyone runs across this data set, I would appreciate knowing about it.

Used in Text: not used
}

\examples{
 str(saving)
}
\keyword{datasets}
