% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twoyear.R
\docType{data}
\name{twoyear}
\alias{twoyear}
\title{twoyear}
\format{A data.frame with 6763 rows and 23 variables:
\describe{
 \item{female}{=1 if female}
 \item{phsrank}{percent high school rank; 100 = best}
 \item{BA}{=1 if Bachelor's degree}
 \item{AA}{=1 if Associate's degree}
 \item{black}{=1 if African-American}
 \item{hispanic}{=1 if Hispanic}
 \item{id}{ID Number}
 \item{exper}{total (actual) work experience}
 \item{jc}{total 2-year credits}
 \item{univ}{total 4-year credits}
 \item{lwage}{log hourly wage}
 \item{stotal}{total standardized test score}
 \item{smcity}{=1 if small city, 1972}
 \item{medcity}{=1 if med. city, 1972}
 \item{submed}{=1 if suburb med. city, 1972}
 \item{lgcity}{=1 if large city, 1972}
 \item{sublg}{=1 if suburb large city, 1972}
 \item{vlgcity}{=1 if very large city, 1972}
 \item{subvlg}{=1 if sub. very lge. city, 1972}
 \item{ne}{=1 if northeast}
 \item{nc}{=1 if north central}
 \item{south}{=1 if south}
 \item{totcoll}{jc + univ}
}}
\source{
\url{http://www.cengage.com/c/introductory-econometrics-a-modern-approach-6e-wooldridge}
}
\usage{
data(twoyear)
}
\description{
Data from wooldRidge package loads lazily. Type data(twoyear) into the console.
}
\examples{
 str(twoyear)
}
\keyword{datasets}
