% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vote1.R
\docType{data}
\name{vote1}
\alias{vote1}
\title{vote1}
\format{A data.frame with 173 rows and 10 variables:
\itemize{
 \item state. state postal code
 \item district. congressional district
 \item democA. =1 if A is democrat
 \item voteA. percent vote for A
 \item expendA. camp. expends. by A, $1000s
 \item expendB. camp. expends. by B, $1000s
 \item prtystrA. percent vote for president
 \item lexpendA. log(expendA)
 \item lexpendB. log(expendB)
 \item shareA. 100*(expendA/(expendA+expendB))
}}
\source{
\url{http://www.cengage.com/cgi-wadsworth/course_products_wp.pl?fid=M20b&product_isbn_issn=9781111531041}
}
\usage{
data(vote1)
}
\description{
Data loads lazily. Type data(vote1) into the console.
}
\examples{
 str(vote1)
}
\keyword{datasets}
