% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wooldRidge.R
\docType{data}
\name{wage2}
\alias{wage2}
\title{wage2}
\format{A data.frame with 935 rows and 17 variables:

\describe{
  \item{wage}{monthly earnings}
  \item{hours}{average weekly hours}
  \item{IQ}{IQ score}
  \item{KWW}{knowledge of world work score}
  \item{educ}{years of education}
  \item{exper}{years of work experience}
  \item{tenure}{years with current employer}
  \item{age}{age in years}
  \item{married}{=1 if married}
  \item{black}{=1 if black}
  \item{south}{=1 if live in south}
  \item{urban}{=1 if live in SMSA}
  \item{sibs}{number of siblings}
  \item{brthord}{birth order}
  \item{meduc}{mother's education}
  \item{feduc}{father's education}
  \item{lwage}{natural log of wage}
}}
\usage{
wage2
}
\description{
wage2
}
\keyword{datasets}
